2023/devops/dj_react_docker_postgres/backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.7.0
click==8.1.6
Django==4.2.3
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.9.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
